Correccion de bug ey implementacion total de envio sms

+++ Lab_RTOS.X/tasks/task_GET_TEMP.h
@@ -7,7 +7,7 @@
 void vTaskTemperature(void * args);
 void Temp_Init() ;
 void setear_leds(ws2812_t);
-void set_temp_umbral(uint8_t n);
+void set_temp_umbral(int n);
